Rename to common name
